# public/openapi.yaml
openapi: 3.1.0
info:
  title: Newman-Server
  version: '1.0'
  description: A simple API to run Postman collections, using the newman engine
paths:
...
  /convert/html:
    post:
      operationId: convertHtml
      summary: Convert newman JSON summary to HTML result
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OutputSummaryFile'
            examples:
              CollectionRunJsonOuputExample:
                $ref: '#/components/examples/CollectionRunJsonOuputExample'
      responses:
        '200':
          description: The HTML reporter output
          content:
            text/html:
              example: HTML output
        '400':
          description: The provided summary file is not of valid format
          content:
            application/json:
              examples:
                application/json:
                  value: |
                    {
                      "errors": "The summary file is not of valid format. Please use the summary generated using Newman."
                    }
components:
  schemas:
...
    OutputSummaryFile:
      type: object
      description: The summary file to run.
      required:
        - summaryFile
      properties:
        summaryFile:
          type: string
          format: binary
          description: the JSON collection run summary file.
  examples:
...
    CollectionRunJsonOuputExample:
      value:
        summaryFile: https://github.com/criteo/newman-server/blob/main/tests/resources/summary-valid.json
servers:
  - url: /
...